+++ crates/lynx-proxy/src/routes/settings/components/GeneralSetting/index.tsx
@@ -1,18 +1,24 @@
 import { LanguageSelector } from '@/components/LanguageSelector';
-import { useGeneralSetting } from '@/store/useGeneralState';
-import { Button, Form, InputNumber, message, Space, Typography } from 'antd';
+import { ConnectType, useGeneralSetting } from '@/store/useGeneralState';
+import { Button, Form, InputNumber, message, Select, Space, Typography, App } from 'antd';
 import React from 'react';
 import { useI18n } from '@/contexts';
 import { CommonCard } from '../CommonCard';
 
-interface IGeneralSettingProps {}
+interface IGeneralSettingProps { }
 
 export const GeneralSetting: React.FC<IGeneralSettingProps> = () => {
   const [form] = Form.useForm();
-  const { maxLogSize, setMaxLogSize } = useGeneralSetting();
+  const { generalSetting, setGeneralSetting } = useGeneralSetting();
   const [messageApi, contextHolder] = message.useMessage();
   const { t } = useI18n();
+  const { language, setLanguage } = useI18n();
+  const { modal } = App.useApp();
 
+
+  const handleLanguageChange = (value: string) => {
+    setLanguage(value as 'en' | 'zh-CN');
+  };
   return (
     <CommonCard
       title={t('settings.general.title')}
@@ -45,18 +51,35 @@ export const GeneralSetting: React.FC<IGeneralSettingProps> = () => {
         className="w-full"
         layout="vertical"
         form={form}
-        initialValues={{
-          maxLogSize: maxLogSize,
-        }}
-        onFinish={async ({ maxLogSize }) => {
-          setMaxLogSize(maxLogSize);
-          messageApi.success(t('settings.general.actions.save'));
+        initialValues={{ ...generalSetting, language }}
+        onFinish={async ({ language, ...value }) => {
+          if (value.connectType !== generalSetting?.connectType) {
+            modal.confirm({
+              title: t('settings.general.connectType.changeConfirm.title'),
+              content: t('settings.general.connectType.changeConfirm.content'),
+              onOk: () => {
+                setGeneralSetting(value);
+                handleLanguageChange(language);
+                messageApi.success(t('settings.general.actions.save'));
+                location.reload();
+              },
+            });
+          } else {
+            setGeneralSetting(value)
+            handleLanguageChange(language);
+            messageApi.success(t('settings.general.actions.save'));
+          }
+
         }}
       >
         <Typography.Title level={5} className="mb-2">
           {t('settings.general.language')}
         </Typography.Title>
-        <LanguageSelector />
+        <Form.Item
+          name="language"
+        >
+          <LanguageSelector />
+        </Form.Item>
         <Typography.Title level={5} className="mb-2">
           {t('settings.general.maxLogSize.title')}
         </Typography.Title>
@@ -81,6 +104,28 @@ export const GeneralSetting: React.FC<IGeneralSettingProps> = () => {
           ]}
         >
           <InputNumber className="w-full" />
+        </Form.Item>
+        <Typography.Title level={5} className="mb-2">
+          {t('settings.general.connectType.title')}
+        </Typography.Title>
+        <Form.Item
+          colon={false}
+          name={'connectType'}
+          rules={[
+            {
+              required: true,
+              message: t('settings.general.connectType.validation.required'),
+            },
+          ]}
+        >
+          <Select>
+            <Select.Option value={ConnectType.ShortPoll}>
+              {t('settings.general.connectType.shortPoll')}
+            </Select.Option>
+            <Select.Option value={ConnectType.SSE}>
+              {t('settings.general.connectType.sse')}
+            </Select.Option>
+          </Select>
         </Form.Item>
       </Form>
     </CommonCard>